CleanUp: Separating crafting.js file to 3 different files and updating imports
* Separating crafting.js file to RecipeData.js TagsData.js and CraftingTableData.js
* updating and cleaning up imports in all files
